Correcting multiplayer visual for Reverse Hands: On and Main Hand: Left settings
Prior to this change a player with the VR setting
Reverse Hands set to on and the Skin settings Main
Hand set to Left, you'd get a scenario where other
players would see this player as having their arms
swapped along with the held tools.

This would lead to a scenario, for example, where the
player holding a sword and shield and swining the sword
would appear to other players as if this player were
attacking with the shield. (It also results in the shield
being held in an odd way on the inside of the arm instead
of the outside of the arm.) This change looks to
primarily correct the multiplayer visuals for this
specific scenario.

For additional context, this functionality seemed
to be in place to compensate for the scenario where
the Reverse Hands is set to On but the Mand Hand is
set to Right. This would be a common scenario where
a left handed player would play Minecraft in the
default right handed mode, but would need to use
the VR Reverse Hands due to handedness being a more
important setting in VR. This would result in the other
players seeing this player with reversed arms, but at least
the tools would show up in the correct hands, such that
the player attacking with a sword actually looks like
them swinging the sword to other players. The swapped
arms and the strange shield holding were just a side effect
of this.

In this change, I've ensured that the scenario for
Reverse Hands: Off, Main Hand: Right works the same as
it does currently, since this is the setting that all
right handed players would use, and is a majority of the
player base.

I've also ensured that the Reverse Hands: On, Main Hand: Right
also works the same way that it did previously, since
it's still likely a common setup for many left handed players.
It looks strange with the arms swapped, but it does give
accurate information about which tool is in which hand.

I've changed it so that the Reverse Hands: On, Main Hand: Left
will not swap the arms. This allows for a superior left handed
experience, since the skin looks correct, the tools are in
the correct hands, and the shield doesn't do strange things.

The fourth scenario, Reverse Hands: Off, Main Hand: Left,
I've left the same as it was before, but I do question
whether or not that is the better choice... It doesn't
swap the arms, so the skin looks nice, but the tools are
held in the wrong hands, which isn't good. Maybe this
is a scenario where the arms should be swapped, which
would make the skin look bad, but at least the tools
would be in the correct hands... That being said,
I don't understand why anyone would actually be setting
up their game in this way, so it's likely that any
additional changes here would just be overkill. So
leaving the functionality as is shouldn't really
cause any issues.